QuickSettings: Add Theme Switcher Tile (TRDS) [2/2]
This adds a QS tile for Slimroms TRDS that works with CM Quicksettings.

(I had to make a lot of changes to get this code to work.  Some of
these changes may be unnecessary or not done the correct way as I
do not know programming.  If you know a better way to implement
this, please let me know.)

Change-Id: I5b081e5269a66f51f2062f4637207c9bc75ba56e